refactor: extract passive deposit to own component
