+++ src/features/deposit/components/DepositForm/index.tsx
@@ -1,20 +1,13 @@
-import {
-  CopyIcon,
-  ExclamationTriangleIcon,
-  InfoCircledIcon,
-} from "@radix-ui/react-icons"
-import { Button, Callout, Flex, Spinner, Text } from "@radix-ui/themes"
+import { ExclamationTriangleIcon } from "@radix-ui/react-icons"
+import { Button, Callout, Flex, Text } from "@radix-ui/themes"
 import { useSelector } from "@xstate/react"
 import clsx from "clsx"
-import { QRCodeSVG } from "qrcode.react"
 import { useEffect, useState } from "react"
-import CopyToClipboard from "react-copy-to-clipboard"
 import { Controller, useFormContext } from "react-hook-form"
 import { assetNetworkAdapter } from "src/utils/adapters"
 import { AssetComboIcon } from "../../../../components/Asset/AssetComboIcon"
 import { EmptyIcon } from "../../../../components/EmptyIcon"
 import { Form } from "../../../../components/Form"
-import { Input } from "../../../../components/Input"
 import type { ModalSelectAssetsPayload } from "../../../../components/Modal/ModalSelectAssets"
 import { NetworkIcon } from "../../../../components/Network/NetworkIcon"
 import { Select } from "../../../../components/Select/Select"
@@ -26,11 +19,11 @@ import type { BaseTokenInfo, UnifiedTokenInfo } from "../../../../types/base"
 import type { ChainType } from "../../../../types/deposit"
 import { BlockchainEnum } from "../../../../types/interfaces"
 import type { SwappableToken } from "../../../../types/swap"
-import { formatTokenValue } from "../../../../utils/format"
 import { isBaseToken, isUnifiedToken } from "../../../../utils/token"
 import { DepositUIMachineContext } from "../DepositUIMachineProvider"
 import { DepositWarning } from "../DepositWarning"
 import { ActiveDeposit } from "./ActiveDeposit"
+import { PassiveDeposit } from "./PassiveDeposit"
 
 // TODO: Temporary disable deposit through POA bridge
 const ENABLE_DEPOSIT_THROUGH_POA_BRIDGE = true
@@ -142,8 +135,6 @@ export const DepositForm = ({ chainType }: { chainType?: ChainType }) => {
     chainType && network && getAvailableDepositRoutes(chainType, network)
   const isActiveDeposit = availableDepositRoutes?.activeDeposit
   const isPassiveDeposit = availableDepositRoutes?.passiveDeposit
-
-  const [isCopied, setIsCopied] = useState(false)
 
   const [preferredDepositOption, setPreferredDepositOption] = useState<
     "active" | "passive"
@@ -258,61 +249,16 @@ export const DepositForm = ({ chainType }: { chainType?: ChainType }) => {
               />
             )}
 
-          {currentDepositOption === "passive" && (
-            <div className="flex flex-col items-stretch">
-              <h2 className="text-[#21201C] text-center text-base font-bold leading-6">
-                Deposit to the address below
-              </h2>
-              <p className="text-[#63635E] text-center text-sm font-normal leading-5 mb-4">
-                {/* biome-ignore lint/nursery/useConsistentCurlyBraces: space is needed here */}
-                Withdraw assets from an exchange to the{" "}
-                {/* biome-ignore lint/nursery/useConsistentCurlyBraces: space is needed here */}
-                {network != null ? networkSelectToLabel[network] : "(empty)"}{" "}
-                address below. Upon confirmation, you will receive your assets
-                on Defuse within minutes.
-              </p>
-              <div className="flex justify-center items-center w-full min-h-[188px] mb-4 rounded-lg bg-[#FDFDFC] border border-[#F1F1F1] p-4">
-                {depositAddress ? (
-                  <QRCodeSVG value={depositAddress} />
-                ) : (
-                  <Spinner loading={true} />
-                )}
-              </div>
-              {depositAddress && (
-                <Input
-                  name="generatedAddress"
-                  value={
-                    depositAddress ? truncateUserAddress(depositAddress) : ""
-                  }
-                  disabled
-                  className="min-h-14 py-0 mb-4 [&>input]:text-base"
-                  slotRight={
-                    <Button
-                      type="button"
-                      size="2"
-                      variant="soft"
-                      className="px-3"
-                      disabled={!depositAddress}
-                    >
-                      <CopyToClipboard
-                        text={depositAddress}
-                        onCopy={() => setIsCopied(true)}
-                      >
-                        <Flex gap="2" align="center">
-                          <Text>{isCopied ? "Copied" : "Copy"}</Text>
-                          <Text asChild>
-                            <CopyIcon height="14" width="14" />
-                          </Text>
-                        </Flex>
-                      </CopyToClipboard>
-                    </Button>
-                  }
-                />
-              )}
-
-              {renderDepositHint(network, minDepositAmount, token)}
-            </div>
-          )}
+          {currentDepositOption === "passive" &&
+            network != null &&
+            derivedToken != null && (
+              <PassiveDeposit
+                network={network}
+                depositAddress={depositAddress}
+                minDepositAmount={minDepositAmount}
+                token={derivedToken}
+              />
+            )}
         </Form>
         {token &&
           network &&
@@ -505,52 +451,4 @@ function NotSupportedDepositRoute() {
       </Callout.Text>
     </Callout.Root>
   )
-}
-
-const networkSelectToLabel: Record<BlockchainEnum, string> = {
-  [BlockchainEnum.NEAR]: "NEAR",
-  [BlockchainEnum.ETHEREUM]: "Ethereum",
-  [BlockchainEnum.BASE]: "Base",
-  [BlockchainEnum.ARBITRUM]: "Arbitrum",
-  [BlockchainEnum.BITCOIN]: "Bitcoin",
-  [BlockchainEnum.SOLANA]: "Solana",
-  [BlockchainEnum.DOGECOIN]: "Dogecoin",
-  [BlockchainEnum.TURBOCHAIN]: "TurboChain",
-  [BlockchainEnum.AURORA]: "Aurora",
-  [BlockchainEnum.XRPLEDGER]: "XRP Ledger",
-}
-
-function renderDepositHint(
-  network: BlockchainEnum | null,
-  minDepositAmount: bigint | null,
-  token: SwappableToken | null
-) {
-  return (
-    <Callout.Root size="1" color="indigo" variant="soft">
-      <Callout.Icon>
-        <InfoCircledIcon />
-      </Callout.Icon>
-
-      {network != null && (
-        <Callout.Text>
-          Make sure to select {networkSelectToLabel[network]} as the deposit
-          network
-        </Callout.Text>
-      )}
-
-      {minDepositAmount != null && minDepositAmount > 1n && token != null && (
-        <Callout.Text>
-          {/* biome-ignore lint/nursery/useConsistentCurlyBraces: space is needed here */}
-          Minimal amount to deposit is{" "}
-          <Text size="1" weight="bold">
-            {formatTokenValue(minDepositAmount, token.decimals)} {token.symbol}
-          </Text>
-        </Callout.Text>
-      )}
-    </Callout.Root>
-  )
-}
-
-function truncateUserAddress(hash: string) {
-  return `${hash.slice(0, 12)}...${hash.slice(-12)}`
 }
